requirements.txt
# PDF Parsing
docling>=2.0.0

# Embeddings
sentence-transformers>=2.2.0

# Vector Database
chromadb>=0.4.0

# Reranking
flashrank>=0.2.0

# Groq API Client
groq>=0.11.0

# CLI & Display
rich>=13.0.0
typer>=0.9.0

# Utilities
Pillow>=10.0.0
python-dotenv>=1.0.0
